chore: move type TText to text component file
