+++ packages/components/src/components/text/text.tsx
@@ -1,7 +1,19 @@
 import classNames from 'classnames';
 import React from 'react';
 import { isEmptyObject } from '@deriv/shared';
-import { TTextProps } from '../types';
+import { TChildren, TStyle } from '../types';
+
+export type TTextProps = {
+    line_height?: string;
+    size?: string;
+    children?: TChildren;
+    color?: string;
+    align?: string;
+    weight?: string;
+    as?: string;
+    className?: string;
+    styles?: TStyle;
+};
 
 const Text = ({ children, size, color, align, weight, line_height, as, className, styles, ...props }: TTextProps) => {
     const class_styles = {
