+++ packages/components/src/components/types/general.types.ts
@@ -3,15 +3,3 @@ export type TStyle = {
 };
 
 export type TChildren = React.ReactNode | React.ReactNode[];
-
-export type TTextProps = {
-    line_height?: string;
-    size?: string;
-    children?: TChildren;
-    color?: string;
-    align?: string;
-    weight?: string;
-    as?: string;
-    className?: string;
-    styles?: TStyle;
-};
